requirements.txt
mysqlclient
pillow
requests